Update to MovingCollectible & ScoreManager

+++ Assets/Assets (Daryl)/Scripts/End_Controller.cs
@@ -54,7 +54,7 @@ namespace Crescendo.InitialCrescendo
             }
             if(currentHighscore < scoreManager.Percentage)
             {
-                PlayerPrefs.SetInt("Score", scoreManager.Percentage);
+                PlayerPrefs.SetFloat("Score", scoreManager.Percentage);
             }
         }
     }
